requirements.txt
django
django_rest_framework
python_decouple